# chore: simplify CloneBatch code